feat(accounts): let admin manage healthworkers, sys-admin add doctors
Account creation was split unevenly: sys-admin could create admins and
healthworkers, admin could create doctors, and nobody could create a
healthworker from the admin shell. Closes the two gaps in #117.

Backend. The account endpoints move out of the sys-admin-only surface
into `routers/accounts.py`, open to both administrative roles. The fixed
`MANAGEABLE_ROLES` constant becomes `manageable_roles(caller)`:

    sys-admin -> admin, healthworker
    admin     -> healthworker

Every endpoint derives its gate from that one helper rather than
re-implementing the check, so an endpoint added later inherits it. The
roster is filtered server-side by the same helper: an admin never
receives admin or sys-admin rows at all, since that list is precisely the
set of accounts worth attacking. Creating a role outside the caller's
reach is 403 `cannot_manage_role`; an unknown role string stays a 422
from the Literal, because "not yours to create" and "no such role" are
different failures.

The boundary matters because an admin who could mint admin accounts, or
reset an existing admin's password, would have promoted itself to
sys-admin. tests/test_account_role_matrix.py walks the whole caller x
target x endpoint grid rather than the cases someone thought to check.

Doctors stay out of this surface for both callers. A doctor is an
accounts row plus a doctors row with its own invite/approve lifecycle;
deleting here would orphan the profile, and disabling here would write
accounts.disabled_at while booking reads doctors.active.

Frontend. The accounts page and the new-doctor page each become a shared
surface parameterised by its navigation targets and creatable roles, so
/admin/healthworkers and /sysadmin/doctors/new reuse them instead of
copying ~500 lines. Sys-admin needs its own doctor route because the role
guard bounces it out of /admin/*. The role column and filter hide
themselves when the roster holds a single role.

Also fixes the admin nav, where the awaiting-approval badge rendered
inside the map with no per-item check (harmless with one tab, wrong with
two) and /admin prefix-matched every admin route.
