+++ frontend/src/components/admin/nav.tsx
@@ -2,12 +2,31 @@
 
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { Stethoscope } from "lucide-react";
+import { Stethoscope, UserCog } from "lucide-react";
 
 import { cn } from "@/lib/cn";
 import { useDoctorSummary } from "@/lib/use-api";
 
-const NAV = [{ href: "/admin", label: "Doctors", Icon: Stethoscope, exact: false }];
+// `/admin` is both the doctors dashboard and the prefix of every other
+// admin route, so a plain startsWith would light both tabs at once. Each
+// entry states its own match instead.
+const NAV = [
+  {
+    href: "/admin",
+    label: "Doctors",
+    Icon: Stethoscope,
+    isActive: (p: string) => p === "/admin" || p.startsWith("/admin/doctors"),
+    // Only this tab carries the awaiting-approval count.
+    badge: true,
+  },
+  {
+    href: "/admin/healthworkers",
+    label: "Health workers",
+    Icon: UserCog,
+    isActive: (p: string) => p.startsWith("/admin/healthworkers"),
+    badge: false,
+  },
+];
 
 export function AdminNav() {
   const pathname = usePathname() ?? "";
@@ -18,8 +37,8 @@ export function AdminNav() {
   return (
     <nav className="border-b border-[var(--border)] bg-[var(--background)]/85 backdrop-blur supports-[backdrop-filter]:bg-[var(--background)]/70">
       <div className="mx-auto flex max-w-7xl items-center gap-1 overflow-x-auto px-6 py-2">
-        {NAV.map(({ href, label, Icon, exact }) => {
-          const active = exact ? pathname === href : pathname.startsWith(href);
+        {NAV.map(({ href, label, Icon, isActive, badge }) => {
+          const active = isActive(pathname);
           return (
             <Link
               key={href}
@@ -39,7 +58,7 @@ export function AdminNav() {
                 )}
               />
               {label}
-              {pending > 0 && (
+              {badge && pending > 0 && (
                 <span
                   title={`${pending} awaiting approval`}
                   className="inline-flex min-w-[1.25rem] items-center justify-center rounded-full bg-sky-100 px-1.5 py-0.5 font-mono text-[10px] text-sky-700"
